refactor(helm): move serveCharts() function to testutil
